[FIXUP] b10c4ac ("ANDROID: ext4: Handle casefolding with encryption")
This adds support for encryption with casefolding.

Since the name on disk is case preserving, and also encrypted, we can no
longer just recompute the hash on the fly. Additionally, to avoid
leaking extra information from the hash of the unencrypted name, we use
siphash via an fscrypt v2 policy.

The hash is stored at the end of the directory entry for all entries
inside of an encrypted and casefolded directory apart from those that
deal with '.' and '..'. This way, the change is backwards compatible
with existing ext4 filesystems.

Signed-off-by: Daniel Rosenberg <drosen@google.com>
Test: Boots, /data/media is case insensitive
Bug: 138322712
Change-Id: I07354e3129aa07d309fbe36c002fee1af718f348
[atndko: This was broken in latest f2fs-stable merge 203ad48e5a0820f0059f74f160ffbddbf2e99d6c ("Merge branch 'aosp-new/upstream-f2fs-stable-linux-4.14.y' of https://source.codeaurora.org/quic/la/kernel/msm-4.14 into R-base").]
Signed-off-by: atndko <z1281552865@gmail.com>
Signed-off-by: Zlatan Radovanovic <zlatan.radovanovic@fet.ba>